feat: config new database and models for location

+++ database/migrations/2026_04_26_123829_create_lokasi_table.php
@@ -4,8 +4,7 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
-{
+return new class extends Migration {
     /**
      * Run the migrations.
      */
@@ -13,10 +12,10 @@ return new class extends Migration
     {
         Schema::create('lokasi', function (Blueprint $table) {
             $table->id();
-            $table->string('nama',100);
+            $table->string('nama', 100);
             $table->decimal('latitude', 11, 8);
             $table->decimal('longitude', 11, 8);
-            $table->double('radius')->default(0);
+            $table->integer('radius')->default(0);
             $table->boolean('is_active')->default(true);
             $table->timestamps();
         });
